+++ neoforge/src/main/java/com/jvn/villagerretaliation/client/interaction/VillagerInteractionScreen.java
@@ -1082,6 +1082,7 @@ public class VillagerInteractionScreen extends Screen implements VillagerInterac
         addHireRoleOption(HiredVillagerRole.BREWING, "recruit.role_brewing");
         addHireRoleOption(HiredVillagerRole.COOK, "recruit.role_cook");
         addHireRoleOption(HiredVillagerRole.SMELTER, "recruit.role_smelter");
+        addHireRoleOption(HiredVillagerRole.COURIER, "recruit.role_courier");
         addHireRoleOption(HiredVillagerRole.ANIMAL_HANDLING, "recruit.role_animal_handling");
         addHireRoleOption(HiredVillagerRole.NITWIT, "recruit.role_nitwit");
         if (canOfferBuilderService()) {
@@ -1145,6 +1146,7 @@ public class VillagerInteractionScreen extends Screen implements VillagerInterac
         addRoleChangeOption(HiredVillagerRole.BREWING, "recruit.role_brewing", VillagerRecruitRequestPayload.Action.SET_ROLE_BREWING);
         addRoleChangeOption(HiredVillagerRole.COOK, "recruit.role_cook", VillagerRecruitRequestPayload.Action.SET_ROLE_COOK);
         addRoleChangeOption(HiredVillagerRole.SMELTER, "recruit.role_smelter", VillagerRecruitRequestPayload.Action.SET_ROLE_SMELTER);
+        addRoleChangeOption(HiredVillagerRole.COURIER, "recruit.role_courier", VillagerRecruitRequestPayload.Action.SET_ROLE_COURIER);
         addRoleChangeOption(HiredVillagerRole.BUILDER, "recruit.role_builder", VillagerRecruitRequestPayload.Action.SET_ROLE_BUILDER);
         addRoleChangeOption(HiredVillagerRole.ANIMAL_HANDLING, "recruit.role_animal_handling", VillagerRecruitRequestPayload.Action.SET_ROLE_ANIMAL_HANDLING);
         addRoleChangeOption(HiredVillagerRole.NITWIT, "recruit.role_nitwit", VillagerRecruitRequestPayload.Action.SET_ROLE_NITWIT);